requirements.txt
diffusers==0.24.0
moviepy==1.0.3
nltk==3.8.1
spacy==3.7.2
accelerate
chardet
opencv-python
transformers
tqdm
Pillow
